+++ src/main/java/org/group_3/Main.java
@@ -1,4 +1,5 @@
 package org.group_3;
+import org.group_3.windows.HelloWindow;
 
 import javax.swing.*;
