+++ src/main/java/org/group_3/windows/CreateWindow.java
@@ -1,0 +1,24 @@
+package org.group_3.windows;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class CreateWindow {
+
+    public static JFrame createWindow(String title){
+        int width = 400;
+        int height = 200;
+        Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
+        int screenWidth = screenSize.width;
+        int screenHeight = screenSize.height;
+
+        JFrame frame = new JFrame(title);
+        frame.setSize(width, height);
+        int x = (screenWidth - width) / 2;
+        int y = (screenHeight - height) / 2;
+        frame.setLocation(x, y);
+
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        return frame;
+    }
+}
